VOICEDB/LANGDB: use 32-bit offsets
This allows for voice and language data files larger than 64kB,
with an upper limit of 2GB.

No voice/language files are expected to be larger than 2GB. If this
is needed in the future, a 64-bit version of the file format can be
created.

+++ src/include/cainteoir/buffer.hpp
@@ -206,20 +206,20 @@ namespace cainteoir
 
 		const char *pstr()
 		{
-			const char *ret = (const char *)first + u16();
+			const char *ret = (const char *)first + u32();
 			if (ret >= (const char *)last)
 				throw std::runtime_error("end of file");
 			return ret;
 		}
 
-		void seek(uint16_t offset)
+		void seek(uint32_t offset)
 		{
 			current = first + offset;
 		}
 
 		bool eof() const { return current >= last; }
 
-		uint16_t offset() const { return current - first; }
+		uint32_t offset() const { return current - first; }
 
 		uint32_t magic() { return ((uint32_t)u8() << 16) | ((uint32_t)u8() << 8) | ((uint32_t)u8()); }
 	private:
